create ts script to scan & extract translation strings
