+++ .gitignore
@@ -77,7 +77,7 @@ src/data/contributors.json
 # These files are generated by `yarn merge-translations` command
 src/intl/*.json
 i18n/locales
-i18n/namespaces
+i18n/merged
 # Auto generated code when gatsby build the site
 src/gatsby-types.d.ts
